Fix attack (#207)
* Fixes

- Fixed problems with Tower prefab (Projectile was not correctly added)
- Fixed FireRange positions (that were altered by the problem commented
above)
- Fixed attacking code (the case that a weapon was going to shoot a
target that was already killed by another weapon was not considered)
- For now: if the target is already dead, the weapon proceeds to shoot
the next target in line (if it has one)

* Attack routine fixed

Bug to fix:
- Previous code didn't considered the fact that the same target could be
attacked by diferent weapons at the same time, therefore, if a target
was killed by one weapon and the others tried to attack an error came
up.

So now, when a Weapon calls Attack, it runs through the list of targets
and gets the first one available to shoot.

Behaviour:
- If the first target in line is already dead (by the weapon itself or
by another weapon) it is removed of the target list. This is done
successively until there is an available target to shoot or the target
list is empty.
- If there is a target to attack, the Projectile is created and shot
towards that target.

Other changes:
- ReceiveDamage doesn't need to return a boolean anymore to notify the
target's death.

To test it:
- You can set the Base Range of each Tower Weapon to 250 (that provides
Fire Ranges big enough so the same Unit collides with at least two
towers at the same time, so they have the same target).

* Check death changed

Method isDead() inside Unit and Building returning this==null has been
replaced by checking the currentHealth in Weapon.